cand_man_api/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==22.1
django-cors-headers==4.4.0